minikine: use dwe-default-time branch
minikine is based on kinesalite. The latest release (v1.11.6) has a bug
that causes messages to be delivered more than once. This pull request updates
the dependency file to pull the code from a fork that includes the fix.
See https://github.com/mhart/kinesalite/pull/36 for more details.

This closes #19.