Kinesis: add application name
Kinesis consumers share state using DynamoDB tables that are named after the
application name. The name was hard-coded and this commit changes that to make
it configurable. If the application is named as `rdss_am_adapter_480_dev`, the
DynamoDB tables provisioned should be the following:

- `rdss_am_adapter_480_dev_checkpoints`
- `rdss_am_adapter_480_dev_clients`
- `rdss_am_adapter_480_dev_metadata`

The default application name is `rdss_am`.

This closes #37.