dep: update metadata format and vendor
This application started using the dep tool in the very early days. Since then,
the metadata format has changed and this commit catches up with thew new format
while it updates the vendored libraries to the latest version available.

The new metadata format is considered stable and will be considered valid and
readable by any future version of the dep tool.