Download JSON Schema files
This commit adds a script that downloads the JSON Schema files of the RDSS API
from the original repository. It requires a personal access token for GitHub's
API (see usage example above). The schemas are deposited in `./hack/schemas`
which is not tracked by this repository.

    $ ./hack/download-schemas.sh 24e2bfdc25eec185119c7761d8a7d328348baba0

The script will return an error if the download failed. If the token argument
is not provided the script will exit with code zero.

The Dockerfiles have been updated to make use of this script as long as the API
token is provided (via build arg `GITHUB_API_TOKEN`). If the download fails the
build fails. If the token is not provided the schemas won't be downloaded and
the adapter will work as usually and the validator will be a no-op.