Add JSON Schema validation support
Introduce a JSON Schema validator that expects to find the schema files in the
directory indicated via `RDSS_ARCHIVEMATICA_ADAPTER_BROKER.SCHEMAS_DIR`. If
undefined the validator will be a no-op. If the schemas fail to load then
application will throw and error and exit.

The messages when invalid are sent to the Invalid Message Queue.