fix tests to work with UUID identifiers
Recent changes have enforced UUIDs as the type for message ids etc, which
broke some of our tests.

Note that the use of real UUIDs for identifiers is not yet accepted in the
Message API specification, so we have to skip decoding the fixtures we now
load from there until this is part of the stable API release (previously we
maintained our own fixtures locally so were able to change the ids to use
UUIDs, but now we are no longer in control of this).