broker: validate headers
With this commit the message package learns how to validate headers. This
includes some special treatment that was required to validate the version
attribute because our gojsonschema library doesn't support patterns so we did
that manually.

`message.New()` now introduces the minimum headers expected including the
new `generator` added in `v2.1.0`.